api-gateway/requirements.txt
# API Gateway Requirements
fastapi==0.115.12
uvicorn==0.34.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
pydantic-settings==2.8.1
python-multipart==0.0.20
